fix: pass expected component size to ffmpeg
