+++ packages/haiku-formats/src/exporters/BaseExporter.ts
@@ -1,10 +1,12 @@
+import HaikuDOMAdapter from '@haiku/core/lib/adapters/dom/HaikuDOMAdapter';
 import {
   BytecodeTimelineProperties,
   HaikuBytecode,
+  ThreeDimensionalLayoutProperty,
 } from '@haiku/core/lib/api';
 
 export default class BaseExporter {
-  constructor (protected bytecode: HaikuBytecode, protected readonly componentFolder: string) {}
+  constructor (protected readonly bytecode: HaikuBytecode, protected readonly componentFolder: string) {}
 
   /**
    * Internal method for visiting every timeline and applying a callback to it.
@@ -29,5 +31,25 @@ export default class BaseExporter {
         callback(timeline, property);
       }
     });
+  }
+
+  protected getComponentSize (): ThreeDimensionalLayoutProperty {
+    const factory = HaikuDOMAdapter(this.bytecode);
+    const component = factory(
+      null,
+      {
+        mixpanel: false,
+        contextMenu: 'disabled',
+        hotEditingMode: true,
+        autoplay: false,
+      },
+    );
+
+    const size = {...component.size};
+
+    // We only want to run migrations and perform auto-sizing. The component can go out of scope now.
+    component.context.destroy();
+
+    return size;
   }
 }
