+++ packages/haiku-formats/src/exporters/video/videoExporter.ts
@@ -10,13 +10,19 @@ export class VideoExporter extends BaseExporter implements ExporterInterface {
    */
   writeToFile (filename: string, framerate: number) {
     const assetPathPattern = path.join(this.componentFolder, `png-${framerate}`, 'frame-%07d.png');
+    const componentSize = this.getComponentSize();
     return new Promise<void>((resolve, reject) => {
       // First make the palette.
       newFfmpegCommand()
         .addInput(assetPathPattern)
         .withInputOptions(['-framerate', `${framerate}`])
         .withVideoCodec('libx264')
-        .withOutputOptions(['-vf', `fps=${framerate}`, '-pix_fmt', 'yuv420p'])
+        .withOutputOptions([
+          '-vf',
+          `fps=${framerate}, scale=${componentSize.x}:${componentSize.y}`,
+          '-pix_fmt',
+          'yuv420p',
+        ])
         .on('error', (stdout, stderr) => {
           reject(stderr);
         })
